UPDATE | VIEWS

+++ drf-10-05/items/views.py
@@ -4,9 +4,18 @@ from rest_framework.response import Response
 from rest_framework import status
 from .models import Item, CustomerOrder, Category, SellRecord
 from drf_yasg.utils import swagger_auto_schema
+from functools import wraps
 
 
+def authenticated_user_required(func):
+    @wraps(func)
+    def wrapper(request, *args, **kwargs):
+        if not request.user.is_authenticated:
+            return Response("Authentication required", status=status.HTTP_401_UNAUTHORIZED)
+        return func(request, *args, **kwargs)
+    return wrapper
 class GetItemsAPIView(APIView):
+    @authenticated_user_required
     def get(self, request):
         item = Item.objects.all()
         serializer = ItemSerializer(item, many=True)
@@ -16,11 +25,12 @@ class GetItemsAPIView(APIView):
         responses={201: 'Created', 400: 'Bad Request'}
     )
     def post(self, request):
-        serializer = ItemSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_authenticated:
+            serializer = ItemSerializer(data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 class UpdateItemApiView(APIView):
@@ -29,24 +39,26 @@ class UpdateItemApiView(APIView):
         responses={200: 'OKAY', 400: 'BAD REQUEST'}
     )
     def put(self, request, pk, *args, **kwargs):
-        instance = Item.objects.get(id=pk)
-        serializer = ItemSerializer(instance=instance, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_authenticated:
+            instance = Item.objects.get(id=pk)
+            serializer = ItemSerializer(instance=instance, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     @swagger_auto_schema(
         request_body=ItemSerializer,
         responses={200: 'OKAY', 400: 'BAD REQUEST'}
     )
     def patch(self, request, pk, *args, **kwargs):
-        instance = Item.objects.get(id=pk)
-        serializer = ItemSerializer(instance=instance, data=request.data, partial=True)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_authenticated:
+            instance = Item.objects.get(id=pk)
+            serializer = ItemSerializer(instance=instance, data=request.data, partial=True)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 class DeleteItemApiView(APIView):
@@ -55,9 +67,10 @@ class DeleteItemApiView(APIView):
         responses={204: 'No Content', 404: 'Not Found'},
     )
     def delete(self, request, pk, *args, **kwargs):
-        try:
-            instance = Item.objects.get(id=pk)
-        except Item.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-        instance.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.user.is_authenticated:
+            try:
+                instance = Item.objects.get(id=pk)
+            except Item.DoesNotExist:
+                return Response(status=status.HTTP_404_NOT_FOUND)
+            instance.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
